(PC-15975) refactor(analytics): move tests into specific folder
